+++ typescript/infra/config/environments/mainnet3/owners.ts
@@ -16,6 +16,7 @@ export const safes: ChainMap<Address> = {
   base: '0xa7ECcdb9Be08178f896c26b7BbD8C3D4E844d9Ba',
   scroll: '0xa7ECcdb9Be08178f896c26b7BbD8C3D4E844d9Ba',
   polygonzkevm: '0xa7ECcdb9Be08178f896c26b7BbD8C3D4E844d9Ba',
+  mantapacific: '0xa7ECcdb9Be08178f896c26b7BbD8C3D4E844d9Ba',
 };
 
 // export const owners = safes;