+++ typescript/helloworld/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @hyperlane-xyz/helloworld
 
+## 3.6.0
+
+### Patch Changes
+
+- Updated dependencies [67a6d971e]
+- Updated dependencies [612d4163a]
+- Updated dependencies [0488ef31d]
+- Updated dependencies [8d8ba3f7a]
+  - @hyperlane-xyz/sdk@3.6.0
+  - @hyperlane-xyz/core@3.6.0
+
+## 3.5.1
+
+### Patch Changes
+
+- Updated dependencies [a04454d6d]
+  - @hyperlane-xyz/sdk@3.5.1
+  - @hyperlane-xyz/core@3.5.1
+
+## 3.5.0
+
+### Patch Changes
+
+- Updated dependencies [655b6a0cd]
+- Updated dependencies [08ba0d32b]
+- Updated dependencies [f7d285e3a]
+  - @hyperlane-xyz/sdk@3.5.0
+  - @hyperlane-xyz/core@3.5.0
+
 ## 3.4.0
 
 ### Patch Changes