+++ typescript/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @hyperlane-xyz/utils
 
+## 3.6.0
+
+## 3.5.1
+
+## 3.5.0
+
 ## 3.4.0
 
 ### Patch Changes